Bump scrapeTimeout up on relevant release exporters
The default scrape timeout for prometheus is 10s. Since we don't scrape
nearly that often, this bumps up the timeout to match the interval. As
long as the scrape returns in less time than the interval now, it will
still work, and the interval can be configured from the values.

This is a hacky solution to allow us to grab metrics for multiple
projects with a single metrics exporter. We encountered this issue
because we added a new project to the github exporter, which took it
over the timeout limit.

A much more scalable solution would be to create separate exporters for
each project, or batch the projects intelligently into groups so each
exporter still returns quickly. Until then, this solution should work.